Update Dockerfile
This appears to be the minimum configuration required to run our Rails
app in the production environment.

We need tzdata for the tzinfo Gem that's required by Active Support.

We're compiling the Rails assets when the image is created and
configured Rails to serve those assets. We might want to use something
like nginx to serve these assets in future but this'll do for now.

We're migrating the database before starting the Rails server in the
docker-entrypoint.sh script. We might want a better solution when we
start running multiple containers but this should be OK for now.