Add basic postcode search
We're validating the postcode, geocoding it and then using that to find
branches within 25 miles.

We've hardcoded the 25 mile range for now.

We've added a small number of suppliers and branches to the seeds data
to help us test locally.

We've added a search controller for the postcode form. We suspect this
might change but it's useful enough for now given that it helps us
deliver the story.

Co-authored-by: Paul Battley <pbattley@gmail.com>