Style /branches
This page layout roughly follows the latest design in Sketch. I've used
the Branch name but dummy data for the rest. There's still quite a bit
of work to do on this page, so I haven't attempted to replicate the
designs perfectly - we're using basic gov.uk pattern library styles,
with a couple of small changes in branches.css.