Display all branches if no postcode parameter is specified
This reinstates some behaviour and examples that were removed in this
commit [1] without explanation. @techbelly were relying on this
behaviour in order to write an end-to-end test for uploading supplier
data into the app.

If it turns out we really need to prevent the branches index page being
viewed with *no* postcode filter, it should be easy enough to reinstate
that behaviour.

[1]: 2d1502e46a3dec7e0e8fd97cc5dc3f9a7d343ee6