Display distance to branch in search results
There's some duplication in the kilometre <-> miles conversions in the
controller and template. We'll address this separately.

We considered calculating the distances we're asserting in our feature
spec but decided on hardcoding them for now.

We've added the `distance` CSS class to the markup to allow us to write
the assertion we want.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>