Remove automatic geocoding of branch postcodes
We've decided to move this step out of the app and into the scripts [1]
which parse the source spreadsheets. This has the advantage that we
don't need to think about background processing for now thus simplifying
the app and infrastructure. If it turns out we need to provide a way to
modify the branch postcode data within the app (e.g. via an admin UI),
we can revisit this decision.

[1]: https://github.com/Crown-Commercial-Service/cmp-supply-teacher-data/