Destroy all suppliers & their branches before each import
This will make it possible for us to update the data in production by
re-uploading via the existing HTTP  endpoint.

It should also allow us to simplify the app further, because we can
remove the logic which determines whether a supplier already exists in
the database. I've removed the relevant examples from the
UploadsController spec and I plan to simplify the associated code in
subsequent commits.