Ask buyers whether they're searching for nominated workers
This introduces the first fork in a buyer's journey. I've been told that
all suppliers have a nominated worker rate so the search results page
continues to show all suppliers (i.e. we don't yet need to worry about
filtering suppliers based on those that have nominated worker rates).

I imagine we'll build on the 'Buyer is not looking for a nominated
worker' scenario when we add support for paying the worker through the
school's payroll.

Note that we have to include the nominated_worker and postcode params
when redirecting to the search form so that we send the user back to the
right place (and not directly back to the nominated worker question).
This is also true of the 'back' link on the branches index page.

Note that I'm not very keen on this implementation - particularly the
logic in the search template - but I'm going to wait until we introduce
more branches to see if it makes it clearer how we should solve the
problem.