Fix regression in branches index view spec
Since this commit [1] the "displays list of all branches" example hasn't
been asserting what was intended: the test would pass if the first
supplier's name only appeared once even though it should appear twice
(once for each of the first & second branches). This happened when the
CSS selector in the assertions was changed so it no longer used the
nth-child pseudo-class and two of the assertions became duplicates:

    expect(page).to have_css('h2.govuk-heading-l', text: /First Supplier/)
    expect(page).to have_css('h2.govuk-heading-l', text: /First Supplier/)

The 2nd of these is effectively redundant, because if the 1st one
passes, so will the 2nd; and if the 1st one fails, so will the 2nd.

[1]: 791c57aa19d3b8035c43099ba46f0e314e31ba33