Display nominated worker rate in supplier search results
A Supplier has many Rates. A Rate has a job type (e.g. 'nominated') and
a mark_up (e.g. 0.1, representing 10%).

Note that we'll need to add something like length of employment to Rates
in future but we don't need that for nominated worker rates: they appear
to be the same irrespective of contract length.

It would appear that Rails now adds an implicit validation for a
belongs_to association. I've added the `supplier` to the `Rate` factory
so that it builds a minimally viable `Rate`.