Add a Location class
`point` is derived from `postcode` so it feels like they could be
encapsulated into a class to allow us to hide away some of the
validation logic and reduce the number of things assigned to the view
from this controller. This commit extracts postcode and point into a
Location class.