Add support for employing workers on school payroll
This commits adds another question, "Do you want to put the worker on
the school's payroll?", that's asked when the buyer isn't looking to
hire a nominated worker. Answering "yes" prompts them to enter their
school's postcode. Answering "no" explains that the system can't be used
to hire workers on agency payroll at the moment.

The search results page (branches/index) can now be reached via the
nominated worker route, or non-nominated worker on school payroll route.
I've added the `school_payroll` param to the `request_params` in the the
BranchesController spec so that we're testing that it's included in the
`back_path` if present.