Convert list of branches to Spreadsheet
This service class takes a list of branches and produces an equivalent
spreadsheet.

After some investigation of the available ruby gems we settled on
`axlsx` for the spreadsheet generation as it is the most
feature-complete gem available for generating spreadsheets, including
support for formulas which we are planning to use to provide some
early-stage calculator features.

Unfortunately there doesn't seem to be a gem available for generating
Open Document Format spreadsheets, which GOV.UK in particular are
pushing for[1]. Axlsx supports generating files in the competing
Office Open XML standard[2].

We are depending on the pre-release version of the gem, as it fixes
the ruby deprecation warnings we see if we run the released
version[3].

We also have to use a separate gem to allow the spreadsheets to be
read in the specs as `axlsx` doesn't support reading. For this we're
using `rubyXL` which of the available read-capable gems had an
interface which supports `StringIO` and therefore avoids writing
temporary files.

[1] https://gdstechnology.blog.gov.uk/2018/04/27/open-document-format-in-government-an-update/
[2] https://en.wikipedia.org/wiki/Office_Open_XML
[3] https://github.com/randym/axlsx/issues/501#issuecomment-365604928

Co-authored-by: Ben Griffiths <bengriffiths@gmail.com>