Accreditation body is no longer part of supplier upload
We're planning to use a more up-to-date data source for supplier
accreditation, but that source can only tell us whether a given supplier
is accredited or not; not which body they are accredited by.

Also we understand that buyers should only see accredited suppliers, so
there's no longer any need to display this data. We plan to exclude
suppliers without accreditation from the supplier upload.