Use simple in-memory cache for Geocoder
This will reduce the number of requests we need to make to the Google
API and will improve performance for repeated queries. Obviously there
will be a cache per process and the cached data will be lost every time
a process terminates (e.g. on deployment), but I think it's probably
still worthwhile.