Avoid a blank white area below our footer
We were seeing a white area below the dark grey footer on pages that
didn't have very much content.

Setting the background-color of the .govuk-template class (which is
added to the html element) ensures that any empty space below the footer
is also dark grey. This appears to be how gov.uk have solved the problem
too.