Move feedback link out of the footer
Alaric's motivation for moving this out of the footer (from the
associated Trello card):

> I added the support address as a separate bar so it's easier to toggle
on and off without altering the internal structure of the footer. Note
that the footer may also contain columns of links at some point. Perhaps
not on CMp, but certainly on other CCS services. ie. i'm being mindful
that the new things i design may want to be used elsewhere.