Add Cognito authorization through auth_controller
AWS Cognito provides OpenID Connect on top of OAuth2 to allow us use it
for authentication in our app.

We're using the omniauth and omniauth-oauth2 Gems to connect to Cognito.

We're not currently verifying the signature of the JWT we receive from
Cognito. We'll add that in a future commit.

The 'GET /auth/cognito/callback' test in the Cognito request spec gives
us some confidence that our Cognito OmniAuth strategy is working as
expected by simulating the OAuth2 response with receive from Cognito.

Notes:

The Auth hash schema (which is the data made available in the
'omniauth.auth' key) is described by Omniauth in
https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema.

We've excluded cognito_spec.rb from RSpec/FilePath because RuboCop wants
us to name the folder omni_auth instead of omniauth.

We set the OmniAuth logger to the Rails logger to avoid it writing
directly to STDOUT.

We've had to add `Dotenv.overload '.env.test' to the rails_helper.  This
is necessary for when we're running the test suite using `rake`.
Without this, the `rake` process loads any development environment
variables first (which might contain real Cognito configuration).