Require users to be signed in to access the app
This replaces the HTTP Basic Auth with authentication provided by
Cognito.

Users have to be signed in, via Cognito, to access all controller
actions except:

* AuthController - this needs to be accessible to users before they're
signed in.

* HomeController#gateway - this is the login page so needs to be
accessible to users before they're signed in.

* FacilitiesManagement::UploadsController - this remains protected by
HTTP Basic Authentication so that we can use cURL to upload data.

* SupplyTeachers::UploadsController - this remains protected by
HTTP Basic Authentication so that we can use cURL to upload data.