Make the footer stick to bottom on short pages
The previous implementation added in
c4753ee9fa486e16b00b714142742b24d24c8c22 meant that the footer always
appeared below the content but would "expand" to the bottom of the page.
This looks odd on pages with little content so we've decided to make the
footer stick to the bottom of the page.

Alaric pointed me at a number of possible approaches[1]. I decided to
use the Flexbox approach[2] as the others required a fixed height
footer, or the use of CSS grids which are more experimental. I believe
the Flexbox approach should work in all browsers we're aiming to
support.

Note that the CSS in the article applies `height: 100%` to the `html`
element. I'm doing that indirectly by applying it to the
`govuk-template` class which is applied to the `html` element.

Note that this implementation requires a container to be added to
everything except the footer. This change is in the application layout
and can be best seen using the `--ignore-all-space` switch.

Note that I'm now setting the background-colour of .govuk-template
(which is applied to the `html` element) to white instead of dark-grey.
The `height: 100vh` (where 100vh is 100% of viewport height) directive
on the `body` meant that anything below the viewport was appearing as
dark grey.

[1]: https://css-tricks.com/couple-takes-sticky-footer/
[2]: https://css-tricks.com/couple-takes-sticky-footer/#article-header-id-3