Add first stab at rendering rates for master vendor MSPs
* I've based the HTML in the template and partial template on the
equivalent files for BranchesController#index. However, I haven't added
any of the BEM CSS classes, because I'm not 100% sure how/whether I
should do that.

* The order in which job types and terms are enumerated is currently not
constrained. However, the natural database order appears to be roughly
the right order, so I'm not too bothered about this right now.

* The code assumes a job type either has three rates associated with it
(one per term), or a single rate (with no term). It looks as if this
holds for the current data which is fairly static, but we should
probably constrain this more so it doesn't cause nasty surprises in the
future.

* I'm not sure I like the naming of the controller action, template and
partial template, but it'll do for now.

* The have_rates matcher in the feature spec is a bit of a horror show,
but I wanted to have a matcher which did not rely on CSS classes,
because I'm not sure what CSS we're going to end up with.