Extract text into a locale file
Having plain strings in the views makes it harder to find the right
place to change the copy, and contributes to tests being brittle. It
would also make it time-consuming to produce a localised version of
the service later.

This commit extracts the "user interface" text into a `en.yml` locale
file. I haven't yet touched the copy-heavy index page or sidebar, and
I haven't made any attempt to change the tests in this commit to rely
on something other than the text as localised.