Add Rate#daily_fee
To model the daily fees that can be/are charged by suppliers under lot 3.

A `Rate` is either a `#percentage_mark_up?` (the default until now) or a
`#daily_fee?`. I've had to remove the NOT NULL constraint on the
rates.mark_up column in the database but have added validation to ensure
that there's either a mark_up _or_ a daily_fee.

I considered renaming and re-using the `mark_up` column to hold both the
mark up rate and the daily fee but decided against it given that they're
different types (float and currency).