Extract govuk_form_group_with_optional_error helper method
To remove some duplication in the templates.

This isn't great. It relies on us knowing that the flash error is about
the form group that we use this helper for. It'd be better if we had
field/attribute specific errors so that we could use something like this
helper for all govuk-form-group elements.