Ensure locale file does not have unused keys
We consider keys under `errors.` to be used because they are referred
to indirectly in PostcodeValidator and LocationValidator and the
`i18n-tasks unused` cannot determine this by static analysis.

We make use of the Faker gem's localisation[1] in our factories so we
also consider all keys under `faker.` to be used.

[1] https://github.com/stympy/faker#customization