Improve error messages
As per the guidance in the GOV.UK Design System[1][2][3].

The i18n-tasks Gem was complaining that the new keys in the en.yml
locale file weren't being used - it only appears to detect explicit
calls to `translate`. We've avoided this problem by adding them to the
`ignore_unused` key in the i18n-tasks.yml config file.

[1]: https://design-system.service.gov.uk/components/error-message/
[2]: https://design-system.service.gov.uk/components/radios/#error-messages
[3]: https://design-system.service.gov.uk/components/text-input/#error-messages

Co-authored-by: James Mead <3169+floehopper@users.noreply.github.com>