No need to pass options to ActiveModel::Errors#add in validators
We're only ever enabling or disabling these validators, so there's never
any options to pass on. And in any case, it's not obvious to me that the
options passed into the `validates` call in a model are appropriate to
be passed into ActiveModel::Errors#add. The only option mentioned in the
documentation for the latter is `strict` which would raise an exception
instead of adding the error and I don't think we're ever going to want
that.