Reset Geocoder::Lookup::Test after examples which setup stubs
The add_stub method is a class method and thus I think the stubs which
are setup in one example may leak into other examples unless the reset
method is called.

The use of before & after blocks in feature specs might be a bit
controversial, but I think it's worthwhile.