Upgrade Ruby v2.5.1 -> v2.5.3 to fix security vulnerabilities
This was originally attempted in this commit [1], but partially rolled
back in this commit [2] due to the lack of an AWS CodeBuild image for
Ruby v2.5.3. A workaround for one of the vulnerabilities was applied in
this commit [3].

The new CMpDevBuildImage_Ruby [4] repo provides a way to build an AWS
CodeBuild image for Ruby v2.5.3 and I believe the CCS infrastructure now
supports using this image.

By upgrading the version of Ruby, this commit fixes CVE-2018-16396
"Tainted flags are not propagated in Array#pack and String#unpack with
some directives" [5] and includes a proper fix for CVE-2018-16395
"OpenSSL::X509::Name equality check does not work correctly" [6],
rendering the workaround in the Gemfile redundant.

Note that I've retained a specific version of Ruby (ruby:2.5.3-alpine)
in the Dockerfile rather than reverting to (ruby:2.5-alpine) to avoid
the same problem we had before where this image which is used for
deployment being updated, but the image used to run the tests not being
updated.

Note also that as per this comment [7], we're now explicitly installing
npm with node.js since apparently the nodejs package on the new alpine
ruby image doesn't include npm.

[1]: 772153d7f3e95fc52c54ae73982c96537eae1ac9
[2]: 5efda83207ba3c1a0fdb85fd60ba04d64c8efc4d
[3]: b227dd057ad5e62e1ad5ed1d83726d42f07a0833
[4]: https://github.com/Crown-Commercial-Service/CMpDevBuildImage_Ruby
[5]: https://www.ruby-lang.org/en/news/2018/10/17/not-propagated-taint-flag-in-some-formats-of-pack-cve-2018-16396/
[6]: https://www.ruby-lang.org/en/news/2018/10/17/openssl-x509-name-equality-check-does-not-work-correctly-cve-2018-16395/
[7]: https://github.com/Crown-Commercial-Service/crown-marketplace/pull/14#issuecomment-434621371