Remove invalid? method from Journey
It was hard in testing to work out whether to stub valid? or invalid?
One of them is defined in terms of the other. This commit solves this
by removing invalid? I wasn't sure which one to remove from Journey,
but most rails examples I've seen use valid? over invalid?