Add data and model for FM regions
These are a mixture of NUTS 2 (for England) and NUTS 3 (for Scotland,
Wales and Northern Ireland) with slightly different names (for
usability, I assume).

Mapping to NUTS 2 and, where appropriate, NUTS 3 region is implemented.

The pseudo entry for "National coverage" is omitted, as this is better
implemented via inference. "International coverage" is retained.