Combine FM suppliers into a single controller action
We'd like to dynamically populate the list of suppliers from the
database. It'll be easier to make this change if we have a single
controller action and view. The current `lot1*_suppliers` views differ
only by the URL they link to (which we will remove when we have the
supplier information in the page), and the presence of some additional
"direct award" guidance for Lot 1a. This commit combines them and
simplifies the preceding code path.