Optimise deletion of FM data on upload
We were destroying the suppliers one by one and relying on the
`dependent: :destroy` option on the supplier's has_many associations to
similarly delete the child objects one by one.  This was very
inefficient. This change halves the time it takes to do the FM upload
from ~60s to ~30s on my laptop.

The downside of this is that we're now bypassing any before_destroy
callbacks that may be added in the future and the
`delete_all_with_dependents` method is effectively duplicating the
`dependent: :destroy` option on the assocations. However, I'm concerned
that the upload HTTP request will timeout in production if we don't make
this change.