Rename TeacherSupplyJourney -> SupplyTeachersJourney
This brings the class name into line with some other bits of the code,
e.g. the supply_teachers route & action.

I don't think either name is perfect, because they're both somewhat
ambiguous, but I think the new name is less surprising in terms of how
people refer to the app.

I'm about to move a bunch of models into a SupplyTeachers namespace and
this discrepancy would have become all the more obvious and confusing.