Add temp-to-perm fee calculator
This calculator implements some of the logic for the calculation in
Schedule 3 of the supply teacher agency framework as explained using
prose and screenshots of Excel in Trello[1].

This calculator does not currently include any logic for capping the
fee based on the notice period, that will have to be implemented
later.

The explanation mentions that bank holidays must be excluded in the
working days calculation, but does not specify which bank holidays to
include where those differ by country. On the basis of other
conversations about sign-in, I've assumed we will use English bank
holidays as provided by the `gb_eng` option in the Holidays gem.

In case it's not obvious to anyone reading this, and it wasn't to me,
the calculation of the daily_supplier_fee gives the agency's
proportion of the mark-up. That is, for example, if you pay £100 and
the markup is 15% then the teacher receives £86.96 and the agency
receives £13.04 (£86.96 * 1.15 = £100).

[1] https://trello.com/c/0DH6ybk5/154-buyers-can-estimate-the-cost-of-hiring-a-temporary-worker-as-permanent#comment-5bd9ab593fd0b289d574289a