Move journey start path from /:journey -> /:journey/start
Previously it was impossible to have the journey home page containing
the start button have the same initial path segment as the journey name.

Thus for Supply Teachers, the journey home page is '/supply-teachers',
but the journey name is 'teacher-supply',  and thus all subsequent ST
paths are of the form '/teacher-supply/*'.

Similarly for Facilities Management, the journey home page is
'/facilities-management', but the journey name is 'fm', and thus all
subsequent FM paths are of the form '/fm/*'.

Making this change will allow us to keep the ST journey home page at
'/supply-teachers', but change the journey name to 'supply-teachers', so
that all subsequent ST path are of the form '/supply-teachers/*'.

Similarly we'll be able to keep the FM journey home page at
'/facilities-management', but change the journey name to
'facilities-management', so that all subsequent paths are of the form
'/facilities-management/*'.

I think this will give us a more consistent set of URL paths.