Add skeleton of management consultancy journey
Adds boiler plate for the management consultancy journey, with
a lot selection page (lacking copy and design!) and an end step
that links to the existing lists of supplier_by_lot on the ccs
agreement finder website.

These pages are roughly drawn and intended to be thrown away when
the supplier data is in the system and being displayed.