Introduce Journey#previous_questions_and_answers
So that we can avoid having to explicitly list the fields to ignore.

Unfortunately this doesn't yet work. There's a problem with (at least) the
`SchoolPostcode` step in that it defines `worker_type` as an attribute,
even though `worker_type` comes from the previous `WorkerType` step. I
wonder whether it'd be useful to differentiate between the attributes
we're going to ask the user to enter, and those that come from previous
steps?

I've made Journey#params a HashWithIndifferentAccess so that I can
slightly simplify the implementation of
`previous_questions_and_answers`. The `params` contain keys as strings
while the `#attributes` method returns an array of symbols.