Split SchoolPostcode step into two
To represent the two paths that can be taken to get to this question.

The advantage is that we no longer have any question steps defining
attributes to represent data that came from previous questions. This
means that my changes in the previous commit now work as expected.

The disadvantage is some additional duplication. And that question
step/attribute relationship is a bit hidden.