Split SchoolPostcode step into two
To represent the two paths that can be taken to get to this question.

I want to introduce a `Journey#previous_questions_and_answers` method
based on the attributes defined in the question steps. This wasn't
possible before this change because the `SchoolPostcode` step included
the `worker_type` attribute that came from a previous question.