MC Suppliers can now be filtered by services offered and lot
Adds a method to suppliers to that takes a lot and list of services
and returns the suppliers that offer the services in the requested
lot. Conceptually, this is similar to the way we implemented
location picking for facilities management, so I use a very similar
query.

Note: Only the services in the lot requested are considered when
choosing a supplier. Rather than throw an exception if services
from across lots are passed in, I filter the service codes list to
include just those in the lot requested.