Implement Early Hire Fee in Term to Perm calculator
The school pays an early hire fee if they hire a contractor within the
first 40 working days of their contract. The fee is equal to the
supplier day rate multipled by the number of working days between the
hire date and 60 working days from the contract start date.

We've updated the feature spec to make it more realistic. The figure
matches the figure in the Excel tool with the same input data where the
agency selected has a markup rate of 30%.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>