Display info about fee for lack of notice in temp to perm calc
From the document explaining the rules for the calculator:

> If a school hires a worker between the 41st working day of a contract
and the 60th working day of a contract, they may be charged both an
early hire fee and a fee for lack of notice, but the total fee
chargeable cannot be greater than 20 working days multiplied by the
working day suppliers fee.

> If a school hires a worker on or after the 61st working day of the
contract, the school may be charged a fee for lack of notice, up to a
maximum of 20 working days multiplied by the working day suppliers fee.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>