Conditionally display Early Hire Fee explanation
From the description of the rules of the Excel tool:

> Schools may be charged an early hire fee if they hire a worker within
the first 40 working days of the contact (equates to 8 working weeks).
The early hire fee is equivalent to the difference in working days
between the proposed hire date and the 60th working day (the last
working day in the 12th working week), multiplied by the working day
supplier fee. Schools may not be charged any additional fee for lack of
notice prior to the 41st working day.

> If a school hires a worker between the 41st working day of a contract
and the 60th working day of a contract, they may be charged both an
early hire fee and a fee for lack of notice, but the total fee
chargeable cannot be greater than 20 working days multiplied by the
working day suppliers fee.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>