Template mailto links spell out whole email address
In testing the site, we found that some users' browsers
were not able to correctly handle mailto links and users
were typically, therefore, unable to continue.

This commit adds an application helper method that builds
mailto links with the anchor text set to the email address.
This change required some tweaks to the beta banner so that
it read well again, and a tweak to the beta banner layout
to span the whole of the column.