Feedback email address no longer comes from ENV
There doesn't seem much value in configuring the feedback email
address in the environment. It will be the same for all envs and
the infrastructure as is doesn't make it any easier to change
an environment variable than it does to deploy.

Because it's now hard-coded in the Marketplace module, we can
remove some logic and tests around checking if it's present.