Use hidden_fields_for_previous_steps_and_responses even in first steps
Even though strictly speaking it's not necessary to call this helper in
templates for the first step in a journey, doing so does no harm and has
the benefit of making the step templates more consistent and easier to
reorder.

I can imagine we might want to extract some of this duplication into a
form builder at some point and doing this will make that easier.