Add DfE Sign-in using omniauth_openid_connect Gem
We're using the omniauth_openid_connect Gem after learning that it's
being used in the Teacher Vacancy Service[1].

DfE Sign-in is disabled by default but can be enabled by setting the
`DFE_SIGNIN_URL` environment variable (along with
`DFE_SIGNIN_CLIENT_ID`, `DFE_SIGNIN_CLIENT_SECRET` and
`DFE_SIGNIN_REDIRECT_URI`). I've had to add dummy values for these
environment variables to .env.test so that DfE Sign-in is enabled in the
tests.

I'm additionally storing the OmniAuth provider in the session so that we
can do provider-specific things in the `AuthController#sign_out` action.
We automatically sign the user out of their Cognito session if that's
how they signed in as we don't expect them to be using Cognito for
anything else. We _don't_ sign the user out of their DfE Sign-in session
as they might be using that to access other services.

Note that I've had to manually set the `provider`s in the
`OmniAuth::AuthHash`s returned in the mock mode. The provider is
populated automatically when using real strategies but apparently not in
the test mode.

[1]: https://github.com/dxw/teacher-vacancy-service/