Redirect requests for DfE callback to DfE sign in
DfE Sign-in will redirect users to our callback endpoint when the user
follows the forgotten password flow after starting the sign in process
from our app. By default this results in the user being presented with a
401 response because the omniauth_openid_connect strategy expects the
callback to be invoked as part of a valid OpenID Connect exchange.

I'm interpreting requests for /auth/dfe/callback with an empty
querystring as those arriving outside of a normal OpenID Connect
exchange (e.g. via the forgot password flow). This mirrors the
implementation in the Teacher Vacancy Service[1][2].

I'm not very keen on using `!OmniAuth.config.test_mode` in the DfeSignIn
middleware. I've added it so that the DfE Sign-in examples in the
authentication feature continue to pass: these examples rely on being
able to get the callback path directly.

[1]: https://github.com/dxw/teacher-vacancy-service/commit/0b4980c0c18f0ee788852f97153a18940be04f90
[2]: https://github.com/dxw/teacher-vacancy-service/commit/b843e94d2122f39b61a3b02a25931fbf4e158c80