Remove references to "Early Hire Fee"
This terminology is not used in the framework document, so may be
confusing to users. It is used in some of the supplementary
explanations we have received to help develop the calculator so I
haven't renamed all the methods that refer to it as I think the
terminology is still useful in that context.