Refactor Calculator#working_days_from_contract_start
To bring it into line with the implementation of `#working_days_before`.

This also avoids a potential problem with the previous implementation in
that it would only work up to a maximum of a year into the future.
That's not a problem for this calculator for now but is an unnecessary
limitation.