Encapsulate login in an authorising object
This retains some of the auth information so that it can say whether the
user is permitted to view a particular resource.

The session store doesn't seem to store complex objects in all
situations (for example, in feature specs), so this is serialised as
JSON in the actual session variable.