Provide supply teacher data download for auditing
To allow stakeholders to easily audit the data we provide the content
of the `supply_teacher_branches`, `supply_teacher_rates` and
`supply_teacher_suppliers` tables in an Excel workbook form to
download at `/supply-teachers/downloads.xlsx`.

I've omitted the id fields from the workbook and denormalised the
supplier name into the other two worksheets to make it easier for the
downloader to cross-reference.

The download is protected by the login mechanism and will redirect to
the gateway page if the downloader is not logged in.