Fix display of validation errors on TTPC ContractStart date fields
Only the error summary was being displayed correctly prior to this
commit. This was mainly because a bunch of methods on ApplicationHelper
was being called with `:contract_start`  which does not correspond to an
attribute on the ContractStart step model. This meant that the helper
methods were never finding any validation errors, because the errors
were being stored against the more specific attribute name keys, e.g.
`:contract_start_day`, `:contract_start_month`, etc.

This commit is not a perfect fix, because it means up to three error
messages can appear above each set of the three number fields
representing the date. However, I think it's a step forward. A better
solution might be to introduce custom validation which sets errors on an
aggregate key, i.e. `:contract_start`.

Note that I've made use of the fact that the value of an
`aria-describedby` attribute can be a space-separated list of IDs [1].

This has also highlighted a couple of things to me which I think we
should address in the near future:

* Duplication in step templates - consider using custom form builder
* Duplication in testing validation in view specs - add helper specs for
ApplicationHelper methods

[1]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-describedby_attribute#Value