Fix display of validation errors on TTPC HireDate date fields
c.f. the previous commit which fixed the same problems for the
ContractStart step.

Only the error summary was being displayed correctly prior to this
commit. This was mainly because a bunch of methods on ApplicationHelper
was being called with `:hire_date`  which does not correspond to an
attribute on the HireDate step model. This meant that the helper
methods were never finding any validation errors, because the errors
were being stored against the more specific attribute name keys, e.g.
`:hire_date_day`, `:hire_date_month`, etc.

This commit is not a perfect fix, because it means up to three error
messages can appear above each set of the three number fields
representing the date. However, I think it's a step forward. A better
solution might be to introduce custom validation which sets errors on an
aggregate key, i.e. `:hire_date`.