Add custom validation error messages for when TTPC fields are blank
This is just a first stab at making these more sensible and compliant
with the GOV.UK Design System [1]. The fact that these were not defined
had become more obvious when all the TTPC steps were combined into a
single page.

Note that this only solves the problem for the "presence" validation.
And I'm starting to wonder whether we should really override the default
error messages i.e. in the `en.errors.format` key.

[1]: https://design-system.service.gov.uk/components/error-message/