Move TTPC into ST journey as another LookingFor answer option
This inserts the temp-to-perm calculator steps into the supply teachers
journey by adding a new answer to the looking-for step. The temp-to-perm
calculator is no longer a top-level journey and has been removed from
the main home page. Some further copy changes are planned to make this
new journey clearer, but I think this commit includes the main
functional changes we need.

Although the diff is quite big, the vast majority of the changes are
moving functionality from the temp_to_perm_calculator namespace to the
supply_teachers namespace.