Add whitelisting for DfE signin to protect sensitive data
We want to enable DfE signin in the cmpdev environment, so we can demo
the DfE signin. However, this environment contains commercially
sensitive data and we're not in control of who has DfE accounts.

Adding this whitelisting by email address means we can enable DfE signin
but restrict access to a known subset of trusted users.

The DFE_SIGNIN_WHITELISTED_EMAIL_ADDRESSES.include environment variable
should be set to a comma-separated list of email addresses.