Remove unused generic translation for location validation message
We've defined specific translations for all the places this is needed.

I'm hoping to remove all of the errors.messages translation strings,
because then I should be able to remove the custom errors.format
translation string which in turn will mean that the default validation
error messages will include the attribute name and will read sensibly,
e.g. "can't be blank" -> "Name can't be blank".