Include fee for lack of notice in calculation
This mirrors the behaviour in the Excel version of this calculator.

If the user enters the date they gave/are planning to give notice to the
supplier then it's considered when calculating the fee the school might
have to pay to the supplier.

I've made up the question and hint text so these will almost certainly
need changing.

Note that I've excluded `context` lines from the LineLength Cop in
rubocop.yml to avoid a warning.