Use default errors.format translation string from ActiveModel
We were overriding the default value of "%{attribute} %{message}" [1] so
that the invalid_postcode & invalid_location error messages were not
prefixed by the attribute name. However, I've recently moved these to
more specific translation keys, e.g. this [2] and this [3], and these
more specific translation strings don't appear to be used via the
errors.format formatting.

The advantage of making this change is that if we forget to add specific
translation strings for validation error messages, then we'll see the
default Rails validation error messages, e.g. "Name can't be blank" vs
"can't be blank". I think this is a lot better.

Ideally we'd somehow fail fast if a translation for a validation error
message is missing, but I haven't yet found a nice way to do this.

[1]: https://github.com/rails/rails/blob/v5.2.1.1/activemodel/lib/active_model/locale/en.yml#L3-L4
[2]: activerecord.errors.models.supply_teachers/branch.attributes.postcode.invalid_postcode
[3]: activemodel.errors.models.supply_teachers/journey/school_postcode_nominated_worker.attributes.location.invalid_location,