Add Calculator#chargeable_working_days_based_on_lack_of_notice
This mirrors the behaviour in the Excel version of this calculator.

The school is required to give the supplier at least 20 working days
notice when they want to make a temporary contractor a permanent member
of staff. If they give the supplier fewer than 20 working days notice
then they may be required to make up the difference to the supplier.