Add Calculator#chargeable_working_days
This mirrors the behaviour in the Excel version of this calculator.

If the school hires a contractor within the first 40 working days (8
weeks) of a contract then they are not charged a lack of notice fee.

If the school hires a contractor after the first 40 working days then
they can be charged an early hire fee (for hiring within 12 weeks) and
a lack of notice fee (for not giving at least 4 weeks notice). However,
this fee cannot exceed 20 working days at the supplier's daily rate for
providing the worker.